+++ aws/flatcar-linux/kubernetes/workers/variables.tf
@@ -119,3 +119,16 @@ variable "node_taints" {
   description = "List of initial node taints"
   default     = []
 }
+
+# unofficial, undocumented, unsupported
+
+variable "arch" {
+  type        = string
+  description = "Container architecture (amd64 or arm64)"
+  default     = "amd64"
+
+  validation {
+    condition     = var.arch == "amd64" || var.arch == "arm64"
+    error_message = "The arch must be amd64 or arm64."
+  }
+}